all loader functions moved to loaders.jsx
